# openapi.yaml
openapi: 3.0.2
info:
  description: Moov Wire implements an HTTP API for creating, parsing, and validating Fedwire messages.
  version: v1
  title: Wire API
  contact:
    url: https://github.com/moov-io/wire
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8088
    description: Local development
...
paths:
...
  /files:
    get:
      tags: ['Wire Files']
      summary: List files
      description: List all Wire files created with the Wire service. These files are not persisted through multiple runs of the service.
      operationId: getWireFiles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
      responses:
        '200':
          description: A list of File objects
          headers:
            X-Total-Count:
              description: The total number of Wire files
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireFiles'
  /files/create:
    post:
      tags: ['Wire Files']
      summary: Create file
      description: >
        Upload a new Wire file, or create one from JSON. When uploading a file, query parameters can be used to
        configure the FedWireMessage validation options. For JSON requests, validation options are set in the 
        request body under fedWireMessage.validateOptions.
      operationId: createWireFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
        - name: skipMandatoryIMAD
          in: query
          description: Optional flag to skip mandatory IMAD validation
          required: false
          schema:
            type: boolean
            default: false
            example: true
        - name: allowMissingSenderSupplied
          in: query
          description: Optional flag to allow SenderSupplied to be nil, which is generally the case in incoming files.
          required: false
          schema:
            type: boolean
            default: false
            example: true
      requestBody:
        description: Content of the Wire file (in json or raw text)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WireFile'
          text/plain:
            schema:
              description: A plaintext FED Wire file
              type: string
              example:
      responses:
        '201':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireFile'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
  /files/{fileID}:
    get:
      tags: ['Wire Files']
      summary: Retrieve file
      description: Get the details of an existing File using the unique File identifier that was returned upon creation.
      operationId: getWireFileByID
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: A File object for the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireFile'
        '404':
          description: A resource with the specified ID was not found
    # post:
    #   tags: ['Wire Files']
    #   summary: Updates the specified FEDWire Message by setting the values of the parameters passed. Any parameters not provided will be left unchanged.
    #   operationId: updateWireFileByID
    #   security:
    #     - bearerAuth: []
    #     - cookieAuth: []
    #   parameters:
    #     - name: X-Request-ID
    #       in: header
    #       description: Optional Request ID allows application developer to trace requests through the system's logs
    #       example: rs4f9915
    #       schema:
    #         type: string
    #     - name: fileID
    #       in: path
    #       description: File ID
    #       required: true
    #       schema:
    #         type: string
    #         example: 3f2d23ee214
    #   requestBody:
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/WireFile'
    #   responses:
    #     '201':
    #       description: A JSON object containing a new File
    #       headers:
    #         Location:
    #           description: The location of the new resource
    #           schema:
    #             type: string
    #             format: uri
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/WireFile'
    #     '400':
    #       description: "Invalid File Header Object"
    #       content:
    #         application/json:
    #           schema:
    #             $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
    delete:
      tags: ['Wire Files']
      summary: Delete file
      description: Permanently delete a File and associated message. It cannot be undone.
      operationId: deleteWireFileByID
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted File.
        '404':
          description: A File with the specified ID was not found.
  /files/{fileID}/contents:
    get:
      tags: ['Wire Files']
      summary: Get file contents
      description: |
        Assembles the existing file, computes sequence numbers and totals. Returns plaintext file.
      operationId: getWireFileContents
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: format
          in: query
          description: Optional file type to get file as fixed length or variable length type
          required: false
          schema:
            type: string
            example: variable
        - name: newline
          in: query
          description: Optional new line flag to have new line or no new line
          required: false
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RawWireFile'
        '404':
          description: A resource with the specified ID was not found
  /files/{fileID}/validate:
    get:
      tags: ['Wire Files']
      summary: Validate file
      description: Validates the existing file. You need only supply the unique File identifier that was returned upon creation.
      operationId: validateWireFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireFile'
        '400':
          description: Validation failed. Check response for errors
        '404':
          description: A resource with the specified ID was not found
  /files/{fileID}/FEDWireMessage:
    post:
      tags: ['Wire Files']
      summary: Add Fedwire message to file
      description: Add a Fedwire Message to the specified file.
      operationId: addFEDWireMessageToFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FEDWireMessage'
      responses:
        '200':
          description: Fedwire Message added to File
        '404':
          description: A resource with the specified ID was not found
components:
  schemas:
    WireFile:
      properties:
        ID:
          type: string
          description: File ID
          example: 3f2d23ee214
        fedWireMessages:
          type: array
          items: 
            $ref: '#/components/schemas/FEDWireMessage'
        validateOptions:
          $ref: '#/components/schemas/ValidateOptions'
      required:
        - fedWireMessages
    WireFiles:
      type: array
      items:
        $ref: '#/components/schemas/WireFile'
    RawWireFile:
      type: string
      description: Plaintext Fedwire file
      example: "{3100}121042882Wells Fargo NA"
    FEDWireMessage:
      properties:
        ID:
          type: string
          description: Fedwire Message ID
          example: 3f2d23ee214
        # messageDisposition, receiptTimeStamp, outputMessageAccountabilityData, errorWire is information Appended by the FEDWire Funds Service
        messageDisposition:
          $ref: '#/components/schemas/MessageDisposition'
        receiptTimeStamp:
          $ref: '#/components/schemas/ReceiptTimeStamp'
        outputMessageAccountabilityData:
          $ref: '#/components/schemas/OutputMessageAccountabilityData'
        errorWire:
          $ref: '#/components/schemas/ErrorWire'
        senderSupplied:
          $ref: '#/components/schemas/SenderSupplied'
        typeSubType:
          $ref: '#/components/schemas/TypeSubType'
        inputMessageAccountabilityData:
          $ref: '#/components/schemas/InputMessageAccountabilityData'
        amount:
          $ref: '#/components/schemas/WireAmount'
        senderDepositoryInstitution:
          $ref: '#/components/schemas/SenderDepositoryInstitution'
        receiverDepositoryInstitution:
          $ref: '#/components/schemas/ReceiverDepositoryInstitution'
        businessFunctionCode:
          $ref: '#/components/schemas/BusinessFunctionCode'
        senderReference:
          $ref: '#/components/schemas/SenderReference'
        previousMessageIdentifier:
          $ref: '#/components/schemas/PreviousMessageIdentifier'
        localInstrument:
          $ref: '#/components/schemas/LocalInstrument'
        paymentNotification:
          $ref: '#/components/schemas/PaymentNotification'
        charges:
          $ref: '#/components/schemas/Charges'
        instructedAmount:
          $ref: '#/components/schemas/InstructedAmount'
        exchangeRate:
          $ref: '#/components/schemas/ExchangeRate'
        beneficiaryIntermediaryFI:
          $ref: '#/components/schemas/FinancialInstitution'
        beneficiaryFI:
          $ref: '#/components/schemas/FinancialInstitution'
        beneficiary:
          $ref: '#/components/schemas/Beneficiary'
        beneficiaryReference:
          $ref: '#/components/schemas/BeneficiaryReference'
        accountDebitedDrawdown:
          $ref: '#/components/schemas/AccountDebitedDrawdown'
        originator:
          $ref: '#/components/schemas/Beneficiary'
        originatorOptionF:
          $ref: '#/components/schemas/OriginatorOptionF'
        originatorFI:
          $ref: '#/components/schemas/FinancialInstitution'
        instructingFI:
          $ref: '#/components/schemas/FinancialInstitution'
        accountCreditedDrawdown:
          $ref: '#/components/schemas/AccountCreditedDrawdown'
        originatorToBeneficiary:
          $ref: '#/components/schemas/OriginatorToBeneficiary'
        fiReceiverFI:
          $ref: '#/components/schemas/FIToFI'
        fiDrawdownDebitAccountAdvice:
          $ref: '#/components/schemas/Advice'
        fiIntermediaryFI:
          $ref: '#/components/schemas/FIToFI'
        fiIntermediaryFIAdvice:
          $ref: '#/components/schemas/Advice'
        fiBeneficiaryFI:
          $ref: '#/components/schemas/FIToFI'
        fiBeneficiaryFIAdvice:
          $ref: '#/components/schemas/Advice'
        fiBeneficiary:
          $ref: '#/components/schemas/FIToFI'
        fiBeneficiaryAdvice:
          $ref: '#/components/schemas/Advice'
        fiPaymentMethodToBeneficiary:
          $ref: '#/components/schemas/FIPaymentMethodToBeneficiary'
        fiAdditionalFIToFI:
          $ref: '#/components/schemas/AdditionalFIToFI'
        currencyInstructedAmount:
          $ref: '#/components/schemas/CurrencyInstructedAmount'
        orderingCustomer:
          $ref: '#/components/schemas/CoverPayment'
        orderingInstitution:
          $ref: '#/components/schemas/CoverPayment'
        intermediaryInstitution:
          $ref: '#/components/schemas/CoverPayment'
        institutionAccount:
          $ref: '#/components/schemas/CoverPayment'
        beneficiaryCustomer:
          $ref: '#/components/schemas/CoverPayment'
        remittance:
          $ref: '#/components/schemas/CoverPayment'
        senderToReceiver:
          $ref: '#/components/schemas/CoverPayment'
        unstructuredAddenda:
          $ref: '#/components/schemas/UnstructuredAddenda'
        relatedRemittance:
          $ref: '#/components/schemas/RelatedRemittance'
        remittanceOriginator:
          $ref: '#/components/schemas/RemittanceOriginator'
        remittanceBeneficiary:
          $ref: '#/components/schemas/RemittanceBeneficiary'
        primaryRemittanceDocument:
          $ref: '#/components/schemas/PrimaryRemittanceDocument'
        actualAmountPaid:
          $ref: '#/components/schemas/RemittanceAmount'
        grossAmountRemittanceDocument:
          $ref: '#/components/schemas/RemittanceAmount'
        amountNegotiatedDiscount:
          $ref: '#/components/schemas/RemittanceAmount'
        adjustment:
          $ref: '#/components/schemas/Adjustment'
        dateRemittanceDocument:
          $ref: '#/components/schemas/DateRemittanceDocument'
        secondaryRemittanceDocument:
          $ref: '#/components/schemas/SecondaryRemittanceDocument'
        remittanceFreeText:
          $ref: '#/components/schemas/RemittanceFreeText'
        serviceMessage:
          $ref: '#/components/schemas/ServiceMessage'
      required:
        - senderSupplied
        - typeSubType
        - inputMessageAccountabilityData
        - amount
        - senderDepositoryInstitution
        - receiverDepositoryInstitution
        - businessFunctionCode
    MessageDisposition:
      properties:
        formatVersion:
          type: string
          minLength: 2
          maxLength: 2
          description: formatVersion identifies the Fedwire message format version 30.
          example: '30'
        testProductionCode:
          type: string
          maxLength: 1
          description: |
            testProductionCode identifies if test or production.

            * `T` - Test
            * `P` - Production
          enum:
            - " "
            - "T"
            - "P"
          example: 'T'
        messageDuplicationCode:
          type: string
          description: |
            MessageDuplicationCode

            * ` ` - Original Message
            * `R` - Retrieval of an original message
            * `P` - Resend
          enum:
            - " "
            - "R"
            - "P"
          example: 'R'
        messageStatusIndicator:
          type: string
          description: |
            MessageStatusIndicator

            Outgoing Messages
            * `0` - In process or Intercepted
            * `2` - Successful with Accounting (Value)
            * `3` - Rejected due to Error Condition
            * `7` - Successful without Accounting (Non-Value)

            Incoming Messages
            * `N` - Successful with Accounting (Value)
            * `S` - Successful without Accounting (Non-Value)
          enum:
            - "0"
            - "2"
            - "3"
            - "7"
            - "N"
            - "S"
          example: '0'
    ReceiptTimeStamp:
      properties:
        receiptDate:
          type: string
          description: |
            ReceiptDate is based on the calendar date. (Format MMDD - M=Month, D=Day)
          format: 'MMDD'
          example: '0401'
        receiptTime:
          type: string
          format: 'HHmm'
          description: |
            ReceiptTime is based on a 24-hour clock, Eastern Time. (Format HHmm - H=Hour, m=Minute)
          example: '1305'
        receiptApplicationIdentification:
          type: string
          maxLength: 4
          description: 'ApplicationIdentification'
          example: 'RB11'
    OutputMessageAccountabilityData:
      properties:
        outputCycleDate:
          type: string
          format: 'CCYYMMDD'
          description: OutputCycleDate (Format CCYYMMDD - C=Century, Y=Year, M=Month, D=Day)
          example: '20190401'
        outputDestinationID:
          type: string
          maxLength: 8
          description: OutputDestinationID
          example: '12345678'
        outputSequenceNumber:
          type: string
          maxLength: 6
          description: OutputSequenceNumber
          example: '000001'
        outputDate:
          type: string
          description: |
            Output Date is based on the calendar date. (Format MMDD - M=Month, D=Day)
          format: "MMDD"
          example: '0401'
        outputTime:
          type: string
          format: "HHmm"
          description: |
            Output Time is based on a 24-hour clock, Eastern Time. (Format HHmm - H=Hour, m=Minute)
          example: '1305'
        outputFRBApplicationIdentification:
          type: string
          maxLength: 4
          description: 'OutputFRBApplicationIdentification'
          example: 'OB11'
    ErrorWire:
       properties:
         errorCategory:
           type: string
           description: |
             ErrorCategory
             * `E` - Data Error
             * `F` - Insufficient Balance
             * `H` - Accountability Error
             * `I` - In Process or Intercepted
             * `W` - Cutoff Hour Error
             * `X` - Duplicate IMAD
           enum:
             - E
             - F
             - H
             - I
             - W
             - X
           example: 'E'
         errorCode:
           type: string
           maxLength: 3
           description: ErrorCode
           example: 'E99'
         errorDescription:
           type: string
           maxLength: 35
           description: ErrorDescription
           example: 'Data Error'
    SenderSupplied:
      properties:
        formatVersion:
          type: string
          maxLength: 2
          description: |
            FormatVersion
            30
          example: '30'
        userRequestCorrelation:
          type: string
          maxLength: 8
          description: UserRequestCorrelation
          example: 'TESTDATA'
        testProductionCode:
          type: string
          description: |
            Identifies if test or production.

            * `T` - Test
            * `P` - Production
          enum:
            - T
            - P
          example: 'T'
        messageDuplicationCode:
          type: string
          description: |
            MessageDuplicationCode

            * ` ` - Original Message
            * `R` - Retrieval of an original message
            * `P` - Resend
          enum:
            - " "
            - R
            - P
          example: 'R'
      required:
        - formatVersion
        - userRequestCorrelation
        - testProductionCode
        - messageDuplicationCode
...
    InputMessageAccountabilityData:
      properties:
        inputCycleDate:
          type: string
          format: 'CCYYMMDD'
          description: |
            InputCycleDate (Format CCYYMMDD - C=Century, Y=Year, M=Month, D=Day)
          example: '20191201'
        inputSource:
          type: string
          minLength: 8
          maxLength: 8
          description: InputSource
          example: 'XYZ ABC '
        inputSequenceNumber:
          type: string
          minLength: 6
          maxLength: 6
          description: InputSequenceNumber
          example: '000001'
      required:
        - inputCycleDate
        - inputSource
        - inputSequenceNumber
    WireAmount:
      description: Maximum amount is a penny less than $10 billion.
      properties:
        amount:
          type: string
          minLength: 12
          maxLength: 12
          description: |
            Amount is 12 numeric digits, right-justified with leading zeros.
            It has an implied decimal point and no commas (e.g., $12,345.67 becomes 000001234567).
            Can be all zeros for subtype 90.
          example: '000000100000'
      required:
        - amount
...
    SenderReference:
      properties:
        senderReference:
          type: string
          maxLength: 16
          description: SenderReference
          example: 'Reference'
    PreviousMessageIdentifier:
      description: The identifier for a previous message.
      properties:
        previousMessageIdentifier:
          type: string
          maxLength: 22
          description: PreviousMessageIdentifier
          example: 'Identifier'
...
    ValidateOptions:
      nullable: true
      properties:
        skipMandatoryIMAD:
          type: boolean
          description: Skip validation of the InputMessageAccountabilityData (IMAD) field
          default: false
          example: true
        allowMissingSenderSupplied:
          type: boolean
          description: Allow FedWireMessage.SenderSupplied to be nil
          default: false
          example: true